tgbot/requirements.txt
certifi==2020.6.20
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
pika==1.1.0
psycopg2-binary==2.8.5
pycparser==2.20
python-telegram-bot==12.8
redis==3.5.3
six==1.15.0
tornado==6.0.4
